feat: move ICreateData to baseSQLWorker
