+++ src/mysql/mySQLWorker.ts
@@ -1,14 +1,12 @@
 import { worker } from '../fillers/monaco-editor-core';
 import { MySQL } from 'dt-sql-parser/dist/parser/mysql';
-import { ICreateData } from '../_.contribution';
-import { BaseSQLWorker } from '../baseSQLWorker';
+import { BaseSQLWorker, ICreateData } from '../baseSQLWorker';
 
 export class MySQLWorker extends BaseSQLWorker {
 	protected _ctx: worker.IWorkerContext;
 	protected parser: MySQL;
-	constructor(ctx: worker.IWorkerContext, _createData: ICreateData) {
-		// CreatedData is not required now.
-		super();
+	constructor(ctx: worker.IWorkerContext, createData: ICreateData) {
+		super(ctx, createData);
 		this._ctx = ctx;
 		this.parser = new MySQL();
 	}
