+++ src/pgsql/PgSQLWorker.ts
@@ -1,13 +1,12 @@
 import { worker } from '../fillers/monaco-editor-core';
 import { PostgreSQL } from 'dt-sql-parser/dist/parser/postgresql';
-import { ICreateData } from '../_.contribution';
-import { BaseSQLWorker } from '../baseSQLWorker';
+import { BaseSQLWorker, ICreateData } from '../baseSQLWorker';
+
 export class PgSQLWorker extends BaseSQLWorker {
 	protected _ctx: worker.IWorkerContext;
 	protected parser: PostgreSQL;
-	constructor(ctx: worker.IWorkerContext, _createData: ICreateData) {
-		// CreatedData is not required now.
-		super();
+	constructor(ctx: worker.IWorkerContext, createData: ICreateData) {
+		super(ctx, createData);
 		this._ctx = ctx;
 		this.parser = new PostgreSQL();
 	}
