+++ src/workerManager.ts
@@ -1,5 +1,5 @@
 import { LanguageServiceDefaults } from './_.contribution';
-import { BaseSQLWorker } from './baseSQLWorker';
+import { BaseSQLWorker, ICreateData } from './baseSQLWorker';
 import { editor, IDisposable, Uri } from './fillers/monaco-editor-core';
 
 const STOP_WHEN_IDLE_FOR = 2 * 60 * 1000; // 2min
@@ -58,7 +58,7 @@ export class WorkerManager<T extends BaseSQLWorker> {
 				// passed in to the create() method
 				createData: {
 					languageId: this._defaults.languageId
-				}
+				} as ICreateData
 			});
 
 			this._client = this._worker.getProxy();
